добавил INotifyPropertyChanged  ненмого навел порадок
