+++ Dz5/ListWorkes/ListWorkes.csproj
@@ -6,8 +6,4 @@
     <UseWPF>true</UseWPF>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="model\" />
-  </ItemGroup>
-
 </Project>
