+++ Dz5/ListWorkes/model/Department.cs
@@ -1,0 +1,53 @@
+﻿using System.Collections.ObjectModel;
+
+
+namespace ListWorkes.model
+{
+   
+        public static class Department
+        {
+            
+
+        #region *** Список Департаментов
+        public static string listDepar = "социологииДоцент;" +
+                  "социологииДоцент;" +
+                  "социологииДоцент;" +
+                  "социологииСтарший преподаватель;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииЛаборант-исследователь;" +
+                  "Департамент социологииПрофессор;" +
+                  "Центр перспективных исследований и разработок в сфере образованияДиректор центра;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииАссистент;" +
+                  "Департамент социологииДоцент;" +
+                  "Учебно-научная социологическая;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииПомощник;" +
+                  "Департамент социологииДоцент;" +
+                  "Учебно-научная социологическая;" +
+                  "Департамент социологииДоцент;" +
+                  "Учебно-научная социологическая;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииПрофессор;" +
+                  "Департамент социологииПрофессор;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииПрофессор;" +
+                  "Департамент социологииМенеджер;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииПрофессор;" +
+                  "Департамент социологииПрофессор;" +
+                  "Департамент социологииСтарший;" +
+                  "Департамент социологииДоцент;" +
+                  "Департамент социологииГлавный;" +
+                  "Департамент социологииСтажер;" +
+                  "Департамент социологииПрофессор;" +
+                  "Департамент социологииДоцент";
+
+#endregion
+
+            }
+      }
+
+
