+++ Dz5/ListWorkes/model/Employee.cs
@@ -1,0 +1,60 @@
+﻿using System.Collections.ObjectModel;
+
+
+namespace ListWorkes.model
+{
+   
+        public static class Employee
+        {
+          
+
+            #region*** Список имен
+          public static  string listWork = "Воеводина Екатерина Владимировна;" +
+                "Грунтовский Иосиф Иосифович;" +
+                "Дудина Ольга Мухаметшевна;" +
+                "Епхиев Олег Муратович;" +
+                "Залесский Петр Карлович;" +
+                "Иванова Яна Сергеевна;" +
+                "Кибакин Михаил Викторович;" +
+                "Киселёва Наталья Ильинична;" +
+                "Кораблин Юрий Алексеевич;" +
+                "Котов Дмитрий Алексеевич;" +
+                "Круглова Елена Леонидовна;" +
+                "Кунижева Диана Анзоровна;" +
+                "Лаамарти Юлия Александровна;" +
+                "Лебедева Светлана Юрьевна;" +
+                "Львов Степан Васильевич;" +
+                "Марков Дмитрий Игоревич;" +
+                "Микрюков Владимир Олегович;" +
+                "Мишин Кирилл Юрьевич;" +
+                "Назаренко Сергей Владимирович;" +
+                "Николаев Александр Александрович;" +
+                "Новиков Алексей Викторович;" +
+                "Носкова Антонина Вячеславовна;" +
+                "Оборский Алексей Юрьевич;" +
+                "Орлова Елена Александровна;" +
+                "Осипова Ольга Степановна;" +
+                "Панова Татьяна Викторовна;" +
+                "Перемибеда Павел Александрович;" +
+                "Письменная Елена Евгеньевна;" +
+                "Примаков Вячеслав Леонидович;" +
+                "Проскурина Александра Сергеевна;" +
+                "Прохода Владимир Анатольевич;" +
+                "Прохорова Ирина Геннадьевна;" +
+                "Проценко Полина Андреевна;" +
+                "Разов Павел Викторович;" +
+                "Райдугин Дмитрий Сергеевич";
+
+            #endregion
+
+           
+
+          
+    }
+
+
+       
+}
+  
+
+
